fix(05.5): bind recovery to issued selection
Track actual selection attempts and issued revisions. Mark missing revisions as unknown in operational history.

+++ crates/x4-bridge/src/operational_correlation.rs
@@ -4,7 +4,7 @@ pub fn scope(section: &str) -> String {
         .map_or_else(String::new, |faction| format!("x4:faction:{faction}:ships"))
 }
 
-pub fn attempt(message: &str) -> u8 {
+pub fn attempt(message: &str) -> u64 {
     message
         .strip_prefix("attempt-")
         .and_then(|value| value.parse().ok())
